+++ lib/four_states.dart
@@ -1,10 +1,48 @@
-sealed class FourStates<T> {}
+sealed class FourStates<T> {
+  const FourStates();
 
-class InitialState<T> implements FourStates<T> {
+  R map<R extends Object?>({
+    required R Function() initialState,
+    required R Function(double? progress) loadingState,
+    required R Function(T successObject) succeedState,
+    required R Function(String? failureMessage, dynamic exception) failedState,
+  }) {
+    final FourStates<T> state = this;
+    return switch (state) {
+      InitialState<T>() => initialState(),
+      LoadingState<T>() => loadingState(state.progress),
+      SucceedState<T>() => succeedState(state.successObject),
+      FailedState<T>() =>
+        failedState(state.failureMessage, state.exceptionObject),
+    };
+  }
+
+  R mayBeMap<R extends Object?>({
+    required R Function() orElse,
+    R Function()? initialState,
+    R Function(double? progress)? loadingState,
+    R Function(T successObject)? succeedState,
+    R Function(String? failureMessage, dynamic exception)? failedState,
+  }) {
+    final FourStates<T> state = this;
+    return switch (state) {
+      InitialState<T>() => initialState == null ? orElse() : initialState(),
+      LoadingState<T>() =>
+        loadingState == null ? orElse() : loadingState(state.progress),
+      SucceedState<T>() =>
+        succeedState == null ? orElse() : succeedState(state.successObject),
+      FailedState<T>() => failedState == null
+          ? orElse()
+          : failedState(state.failureMessage, state.exceptionObject),
+    };
+  }
+}
+
+class InitialState<T> extends FourStates<T> {
   const InitialState();
 }
 
-class LoadingState<T> implements FourStates<T> {
+class LoadingState<T> extends FourStates<T> {
   const LoadingState([this.progress]);
 
   final double? progress;
@@ -20,7 +58,7 @@ class LoadingState<T> implements FourStates<T> {
   int get hashCode => progress.hashCode;
 }
 
-class SucceedState<T> implements FourStates<T> {
+class SucceedState<T> extends FourStates<T> {
   const SucceedState(this.successObject);
   final T successObject;
 
@@ -35,7 +73,7 @@ class SucceedState<T> implements FourStates<T> {
   int get hashCode => successObject.hashCode;
 }
 
-class FailedState<T> implements FourStates<T> {
+class FailedState<T> extends FourStates<T> {
   final String? failureMessage;
   final dynamic exceptionObject;